WLAN - Arpspoof + Netdiscover
This is an automated script to simplify and speed up the process of Arpspoofing through disovering hosts within the network through Netdiscover. The process is fairly simple and launches two terminals; one for netdiscover and one for the arpspoof.

TIP: Please be careful when running netdiscover for too long. Too many arprequests may kick you off the network.